Corrected code review comments: code formattig, class naming suggestion..
