+++ core-java/src/main/java/com/baeldung/algorithms/annealing/City.java
@@ -2,35 +2,21 @@ package com.baeldung.algorithms.annealing;
 
 import lombok.Data;
 
-@Data public class City {
+@Data
+public class City {
 
-        private int x;
-        private int y;
+    private int x;
+    private int y;
 
-        public City() {
-                this.x = (int) (Math.random() * 500);
-                this.y = (int) (Math.random() * 500);
-        }
+    public City() {
+        this.x = (int) (Math.random() * 500);
+        this.y = (int) (Math.random() * 500);
+    }
 
-        public double distanceToCity(City city) {
-                int x = Math.abs(getX() - city.getX());
-                int y = Math.abs(getY() - city.getY());
-                return Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2));
-        }
+    public double distanceToCity(City city) {
+        int x = Math.abs(getX() - city.getX());
+        int y = Math.abs(getY() - city.getY());
+        return Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2));
+    }
 
-        public int getX() {
-                return x;
-        }
-
-        public void setX(int x) {
-                this.x = x;
-        }
-
-        public int getY() {
-                return y;
-        }
-
-        public void setY(int y) {
-                this.y = y;
-        }
 }
